requirements.txt
celery==4.2.0
coreapi==2.3.3
ddtrace==0.16.0
Django==2.2
django-admin-rangefilter==0.3.7
django-extensions==2.1.5
django-filter==2.0.0
django-import-export==1.0.1
django-model-utils==3.1.2
django-reversion==3.0.0
djangorestframework==3.8.2
djangorestframework-gis==0.13
drf-url-filters==0.5.1
elasticsearch==6.3.0
gunicorn==19.8.1
marshmallow==2.15.3
raven==6.9.0
redis==2.10.6
splitio-client==6.2.2
python-decouple==3.3
Unipath==1.1
django-cleanup==5.1.0
dj-database-url
django-cors-headers==3.5